% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpvscrape_package.R
\docType{package}
\name{hpvscrape-package}
\alias{hpvscrape}
\alias{hpvscrape-package}
\title{Package hpvscrape}
\description{
hpvscrape is an R library to pull hpv test results from pdfs and return as dataframes
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iluvdata/hpvscrap}
  \item Report bugs at \url{https://github.com/iluvdata/hpvscrape/issues}
}

}
\keyword{internal}
