% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdftools.R
\name{PDFTool}
\alias{PDFTool}
\title{R6 Object to open and scrape PDF Files}
\description{
Scrape text from PDF file, and manage that file
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PDFTool-new}{\code{PDFTool$new()}}
\item \href{#method-PDFTool-getResults}{\code{PDFTool$getResults()}}
\item \href{#method-PDFTool-getPDFs}{\code{PDFTool$getPDFs()}}
\item \href{#method-PDFTool-clone}{\code{PDFTool$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFTool-new"></a>}}
\if{latex}{\out{\hypertarget{method-PDFTool-new}{}}}
\subsection{Method \code{new()}}{
Open pdf file for parsing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFTool$new(pdf, start, end)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pdf}}{a path to pdf or \link{raw} vector containing the pdf of test results}

\item{\code{start}}{a S3 \code{pdf_start} object or character for start of result}

\item{\code{end}}{a S3 \code{pdf_end} object or character for start of result}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{PDFTool} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFTool-getResults"></a>}}
\if{latex}{\out{\hypertarget{method-PDFTool-getResults}{}}}
\subsection{Method \code{getResults()}}{
Vector of results (combined)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFTool$getResults()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
vector of combined results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFTool-getPDFs"></a>}}
\if{latex}{\out{\hypertarget{method-PDFTool-getPDFs}{}}}
\subsection{Method \code{getPDFs()}}{
get specific pdf for a result
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFTool$getPDFs(resNum, output)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resNum}}{integer of the result (index of \code{getResult()})}

\item{\code{output}}{path to where pdf will be written}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
path to pdf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFTool-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PDFTool-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFTool$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
