% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpert.R
\name{TZ}
\alias{TZ}
\title{Time Zone R6}
\description{
The class used to mapping machine name to time zome
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TZ-new}{\code{TZ$new()}}
\item \href{#method-TZ-getTZ}{\code{TZ$getTZ()}}
\item \href{#method-TZ-clone}{\code{TZ$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TZ-new"></a>}}
\if{latex}{\out{\hypertarget{method-TZ-new}{}}}
\subsection{Method \code{new()}}{
Create a \code{TZ} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TZ$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{either a character of a single time zone or mapped time zones i.e.
\code{TZ$new("110016067" = "America/Mexico_City", "110016068" = "America/Sao_Paulo")} where the names are the literal
Xpert machine names on the generated reports and the time zones are one of \code{\link[=OlsonNames]{OlsonNames()}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a TZ object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TZ-getTZ"></a>}}
\if{latex}{\out{\hypertarget{method-TZ-getTZ}{}}}
\subsection{Method \code{getTZ()}}{
Return a time zone string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TZ$getTZ(device_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{device_names}}{the literal machine names for which you want to lookup a timezones.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a time zone stirng
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TZ-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TZ-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TZ$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
