% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpert.R
\name{Xpert}
\alias{Xpert}
\title{Xpert Parsing R6}
\description{
The class used to parse Xpert HPV results
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Xpert-new}{\code{Xpert$new()}}
\item \href{#method-Xpert-getResults}{\code{Xpert$getResults()}}
\item \href{#method-Xpert-getPDFs}{\code{Xpert$getPDFs()}}
\item \href{#method-Xpert-clone}{\code{Xpert$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Xpert-new"></a>}}
\if{latex}{\out{\hypertarget{method-Xpert-new}{}}}
\subsection{Method \code{new()}}{
Parse new pdf of xpert result(s)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Xpert$new(pdf, xpertTZ = Sys.timezone())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pdf}}{path to pdf or \link{raw} vector of pdf}

\item{\code{xpertTZ}}{timezone of xpert machine (see \code{\link[=OlsonNames]{OlsonNames()}}) or \link{TZ} R6 object of machine mapped timezimes}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{Xpert} R6 Object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Xpert-getResults"></a>}}
\if{latex}{\out{\hypertarget{method-Xpert-getResults}{}}}
\subsection{Method \code{getResults()}}{
Get \code{xpert_result} (a \code{data.frame}) for a particular index
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Xpert$getResults(idx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{idx}}{\code{integer} vector of indices of xpert results to return, if not provided, function will return all parsed results}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{xpert_result} class (a specialized \code{data.frame})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Xpert-getPDFs"></a>}}
\if{latex}{\out{\hypertarget{method-Xpert-getPDFs}{}}}
\subsection{Method \code{getPDFs()}}{
Write the single xpert result pdf to disk
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Xpert$getPDFs(idx, output)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{idx}}{single \code{integer} of results parsed in this class}

\item{\code{output}}{path to output pdf (you must specify extension \code{.pdf} yourself). If not provided will create a file named using \code{sample_id} in a temp directory}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the path of the saved pdf
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Xpert-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Xpert-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Xpert$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
